Only get metrics for pods that are scheduled
The join I was using for the GPU metric was sometimes causing issues due to
having many-to-many match which is not allowed. So I have switched to using
"unless". This works reliably. I also applied this to the CPU and memory
metric.

I have also changed the delimiter to a comma because google sheets does not
auto detect pipe.