Bunch of major changes
openshift_prometheus_metrics will now collect and dump the raw metrics of each
kind. There's no use of chunk_dates method anymore as the idea is to collect
small chunk of raw data with that anway.

Then once you have raw metrics for all periods of a month, you will pass that
to merge.py and that will merge and condense the meitrcs, and will ultimately
publish the csv after doing whatever math is needed.