Implement how we gather GPU types.
Ths cluster has multiple types of GPUs. To determine the type of GPUs we
query the node labels where a pod requesting GPU was run, from that we
can find exact model of the GPU.

This commit updates the query that to get the node labels when we gather GPU
usage. The merge metrics method is updated to keep the gpu_type. The GPU type
for lenovo A100s is added and a corresponding SU is also created.