Add mechanism for handling nvidia MIG
We now use 2 parameters, gpu_type and gpu_resource, to figure out what
type of GPU was in use and if it was a MIG GPU.
For a non-MIG gpu, the gpu_resource is `nvidia.com/gpu`, while for a MIG
GPU it's of the form `nvidia.com/mig-Xg.Ygb` where X and Y could change.

Since we don't have service unit definitions for various MIG GPUs, for now
I am calling it "OpenShift Unknown MIG GPU". We will update this once it's
clear what type of SUs we will have.