Add classes to model pod metrics and project invoices.
The class `Pod` represent a pod's metrics and exposes methods to get the
service unit and runtime of a pod. The service unit information is returned as
a namedtuples. The method to get runtime will allow us to ignore certain
hours, though, that will be added in a later commit.

The class `ProjectInvoice` represents invoice data for a project. It has a
method that takes in a pod and aggregates it's usage data. Another method
`generate_invoice_rows` will return the csv data to be exported.

The caller write_metrics_by_namespace has been updated to use these new classes.

Tests have been updated to work with the new changes. There's some duplicate
constants that I will sort in the next few commits.