Update get_service_unit and write_metrics_by_pod.
I updated get_service_unit so it's not a static method anhmore. I don't have
to pass any arguments to it anymore. I can simply call it on the pod object.

That change then broke write_metrics_by_pod because it was making a call to
get_service_unit, so I updated that method to use the new pod class. In that
process I ended up writing a method to generate the pod metric rows so now
write_metrics_by_pod uses the pod class for generating the report. I decided
to remove certain columns for this report, specifically the coldfront attributes
because they are not particularly useful in a pod report.