Accept absolute paths for filenames
For the merge.py, we now take in the names for the invoice file and the pod
report file. The names of the S3 keys are calculated are not affected by this.

In the metrics collection script, the now output file name is also absolute.
We no longer create any directory to organize the metrics in a folder. However
when the names are uploaded to the S3 bucket, the key is organized and the names
are calculated.

This allows us to generated the files in /tmp (or any other directory) with
arbritary names but the file names uploaded to s3 are still calculated.