Break the GPU requests queries into two
The original query creates a join on 2 queries so that the node labels
with the GPU product information are included in the GPU resource request
metrics. The problem is that this will sometimes fail if the kube_node_labels
query returns multiple results for the same node: one series has the labels
put by the GPU operator while the others are missing those.

This commit breaks down the GPU requests query into 2 parts:
1. The first query gets the gpu resource requests for scheduled pods.
2. The second query gets all node labels where the gpu label isn't empty.

We then manually insert the labes into the metrics data.

closes https://github.com/CCI-MOC/openshift-usage-scripts/issues/87