Add query to get pod labels
We specifically get pod labels where the class label isn't empty. This label
is currently only expected to be found for the rhods-notebooks namespace.

Once we have the pod labels, we insert this label with the cpu metrics. This
is arbritary, it could have been merged with memory metrics.

Finally, when we are merging our metrics we add this label to the merged_data
dictionary which is ultimately used for processing.